+++ src/main/java/society/IHuman.java
@@ -6,5 +6,5 @@ public interface IHuman {
     void paint(Graphics graphics);
     int getHealthStatus();
     void setHealthStatus(int healthStatus);
-    int getAntibodies();
+    double getAntibodies();
 }
